+++ Solution Items/RELEASENOTES.md
@@ -5,6 +5,7 @@
 - Replaced current file header with a Yaml Front Matter (breaking!!)
 - Removed MenuBuilder (this can be done via templating now)
 - Available variables for templating changed, please see documentation
+- If there is no command argument given for the site path, the current path is used
 
 **1.0.1**